+++ deployments/liqo/templates/liqo-ipam-deployment.yaml
@@ -49,18 +49,10 @@ spec:
               port: 6000
           {{- end }}
           args:
-            - --pod-cidr={{ .Values.ipam.podCIDR }}
-            - --service-cidr={{ .Values.ipam.serviceCIDR }}
+            - --pod-name=$(POD_NAME)
+            - --port=6000
             {{- if $ha }}
-            - --lease-enabled=true
-            {{- end }}
-            {{- if .Values.ipam.reservedSubnets }}
-            {{- $d := dict "commandName" "--reserved-pools" "list" .Values.ipam.reservedSubnets }}
-            {{- include "liqo.concatenateList" $d | nindent 12 }}
-            {{- end }}
-            {{- if .Values.ipam.additionalPools }}
-            {{- $d := dict "commandName" "--additional-pools" "list" .Values.ipam.additionalPools }}
-            {{- include "liqo.concatenateList" $d | nindent 12 }}
+            - --leader-election
             {{- end }}
             {{- if .Values.common.extraArgs }}
             {{- toYaml .Values.common.extraArgs | nindent 12 }}
@@ -77,8 +69,6 @@ spec:
             valueFrom:
              fieldRef:
                fieldPath: metadata.namespace
-          - name: DEPLOYMENT_NAME
-            value: {{ include "liqo.prefixedName" $ipamConfig }}
           resources: {{- toYaml .Values.ipam.internal.pod.resources | nindent 12 }}
       {{- if ((.Values.common).nodeSelector) }}
       nodeSelector:
