+++ pkg/ipam/ipam.proto
@@ -1,76 +1,62 @@
 syntax="proto3";
 option go_package = "./ipam";
 
-service ipam {
-    rpc MapEndpointIP (MapRequest) returns (MapResponse);
-    rpc UnmapEndpointIP (UnmapRequest) returns (UnmapResponse);
-    rpc MapNetworkCIDR (MapCIDRRequest) returns (MapCIDRResponse);
-    rpc UnmapNetworkCIDR (UnmapCIDRRequest) returns (UnmapCIDRResponse);
-    rpc GetHomePodIP (GetHomePodIPRequest) returns (GetHomePodIPResponse);
-    rpc BelongsToPodCIDR (BelongsRequest) returns (BelongsResponse);
-    rpc GetOrSetExternalCIDR(GetOrSetExtCIDRRequest) returns (GetOrSetExtCIDRResponse);
-    rpc SetSubnetsPerCluster(SetSubnetsPerClusterRequest) returns (SetSubnetsPerClusterResponse);
+service IPAM {
+    rpc IPAcquire (IPAcquireRequest) returns (IPAcquireResponse);
+    rpc IPRelease (IPReleaseRequest) returns (IPReleaseResponse);
+
+    rpc NetworkAcquire (NetworkAcquireRequest) returns (NetworkAcquireResponse);
+    rpc NetworkRelease (NetworkReleaseRequest) returns (NetworkReleaseResponse);
+    rpc NetworkIsAvailable (NetworkAvailableRequest) returns (NetworkAvailableResponse);
 }
 
-message MapRequest {
-    string clusterID = 1;
-    string ip = 2;
+message ResponseResult {
+    bool success = 1;
+    string error = 2;
 }
 
-message MapResponse {
+message IPAcquireRequest {
     string ip = 1;
+    string cidr = 2;
 }
 
-message UnmapRequest {
-    string clusterID = 1;
+message IPAcquireResponse {
+    ResponseResult result = 1;
     string ip = 2;
 }
 
-message UnmapResponse {}
-
-message MapCIDRRequest {
-    string cidr = 1;
-}
-
-message MapCIDRResponse {
-    string cidr = 1;
-}
-
-message UnmapCIDRRequest {
-    string cidr = 1;
-}
-
-message UnmapCIDRResponse {}
-
-message GetHomePodIPRequest {
-    string clusterID = 1;
-    string ip = 2;
-}
-
-message GetHomePodIPResponse {
-    string homeIP = 1;
-}
-
-message BelongsRequest {
+message IPReleaseRequest {
     string ip = 1;
+    string cidr = 2;
 }
 
-message BelongsResponse {
-    bool belongs = 1;
+message IPReleaseResponse {
+    ResponseResult result = 1;
 }
 
-message GetOrSetExtCIDRRequest {
-    string desiredExtCIDR = 1;
+message NetworkAcquireRequest {
+    string cidr = 1;
+    bool immutable = 2; // If true, the network cannot be remapped. It will be allocated if available, or an error will be returned.
 }
 
-message GetOrSetExtCIDRResponse {
-    string remappedExtCIDR = 1;
+message NetworkAcquireResponse {
+    ResponseResult result = 1;
+    string cidr = 2;
 }
 
-message SetSubnetsPerClusterRequest {
-    string remappedPodCIDR = 1;
-    string remappedExternalCIDR = 2;
-    string clusterID = 3;
+message NetworkReleaseRequest {
+    string cidr = 1;
 }
 
-message SetSubnetsPerClusterResponse {}
+message NetworkReleaseResponse {
+    ResponseResult result = 1;
+}
+
+message NetworkAvailableRequest {
+    string cidr = 1;
+}
+
+message NetworkAvailableResponse {
+    ResponseResult result = 1;
+    bool available = 2;
+}
