+++ pkg/ipamold/ipam.proto
@@ -1,0 +1,76 @@
+syntax="proto3";
+option go_package = "./ipam";
+
+service ipam {
+    rpc MapEndpointIP (MapRequest) returns (MapResponse);
+    rpc UnmapEndpointIP (UnmapRequest) returns (UnmapResponse);
+    rpc MapNetworkCIDR (MapCIDRRequest) returns (MapCIDRResponse);
+    rpc UnmapNetworkCIDR (UnmapCIDRRequest) returns (UnmapCIDRResponse);
+    rpc GetHomePodIP (GetHomePodIPRequest) returns (GetHomePodIPResponse);
+    rpc BelongsToPodCIDR (BelongsRequest) returns (BelongsResponse);
+    rpc GetOrSetExternalCIDR(GetOrSetExtCIDRRequest) returns (GetOrSetExtCIDRResponse);
+    rpc SetSubnetsPerCluster(SetSubnetsPerClusterRequest) returns (SetSubnetsPerClusterResponse);
+}
+
+message MapRequest {
+    string clusterID = 1;
+    string ip = 2;
+}
+
+message MapResponse {
+    string ip = 1;
+}
+
+message UnmapRequest {
+    string clusterID = 1;
+    string ip = 2;
+}
+
+message UnmapResponse {}
+
+message MapCIDRRequest {
+    string cidr = 1;
+}
+
+message MapCIDRResponse {
+    string cidr = 1;
+}
+
+message UnmapCIDRRequest {
+    string cidr = 1;
+}
+
+message UnmapCIDRResponse {}
+
+message GetHomePodIPRequest {
+    string clusterID = 1;
+    string ip = 2;
+}
+
+message GetHomePodIPResponse {
+    string homeIP = 1;
+}
+
+message BelongsRequest {
+    string ip = 1;
+}
+
+message BelongsResponse {
+    bool belongs = 1;
+}
+
+message GetOrSetExtCIDRRequest {
+    string desiredExtCIDR = 1;
+}
+
+message GetOrSetExtCIDRResponse {
+    string remappedExtCIDR = 1;
+}
+
+message SetSubnetsPerClusterRequest {
+    string remappedPodCIDR = 1;
+    string remappedExternalCIDR = 2;
+    string clusterID = 3;
+}
+
+message SetSubnetsPerClusterResponse {}
