+++ pkg/ipamold/utils/utils_test.go
@@ -20,7 +20,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	ipamutils "github.com/liqotech/liqo/pkg/ipam/utils"
+	ipamutils "github.com/liqotech/liqo/pkg/ipamold/utils"
 )
 
 const (
