+++ pkg/liqo-controller-manager/networking/external-network/configuration/configuration_controller.go
@@ -28,7 +28,7 @@ import (
 	ipamv1alpha1 "github.com/liqotech/liqo/apis/ipam/v1alpha1"
 	networkingv1beta1 "github.com/liqotech/liqo/apis/networking/v1beta1"
 	"github.com/liqotech/liqo/pkg/consts"
-	"github.com/liqotech/liqo/pkg/ipam"
+	ipam "github.com/liqotech/liqo/pkg/ipamold"
 	"github.com/liqotech/liqo/pkg/utils/events"
 	ipamutils "github.com/liqotech/liqo/pkg/utils/ipam"
 )
